+++ tests/test_main.py
@@ -19,8 +19,6 @@ def pythonVersionBelow310():
 expectedViolations_True = [
     'DOC101: Method `MyClass.func1_3`: Docstring contains fewer arguments than in '
     'function signature. ',
-    'DOC109: Method `MyClass.func1_3`: The option `--arg-type-hints-in-docstring` is `True` '
-    'but there are no type hints in the docstring arg list ',
     'DOC103: Method `MyClass.func1_3`: Docstring arguments are different from '
     'function arguments. (Or could be other formatting issues: '
     'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). Arguments in the function signature but not in '
@@ -55,8 +53,6 @@ expectedViolations_True = [
     'do not match: arg1, arg2',
     'DOC101: Function `func72`: Docstring contains fewer arguments than in '
     'function signature. ',
-    'DOC109: Function `func72`: The option `--arg-type-hints-in-docstring` is `True` '
-    'but there are no type hints in the docstring arg list ',
     'DOC103: Function `func72`: Docstring arguments are different from function '
     'arguments. (Or could be other formatting issues: '
     'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). Arguments in the function signature but not in the '
@@ -66,8 +62,6 @@ expectedViolations_True = [
 expectedViolations_False = [
     'DOC101: Method `MyClass.func1_3`: Docstring contains fewer arguments than in '
     'function signature. ',
-    'DOC109: Method `MyClass.func1_3`: The option `--arg-type-hints-in-docstring` is `True` '
-    'but there are no type hints in the docstring arg list ',
     'DOC103: Method `MyClass.func1_3`: Docstring arguments are different from '
     'function arguments. (Or could be other formatting issues: '
     'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). Arguments in the function signature but not in '
@@ -98,8 +92,6 @@ expectedViolations_False = [
     'these args do not match: arg1, arg2',
     'DOC101: Function `func72`: Docstring contains fewer arguments than in '
     'function signature. ',
-    'DOC109: Function `func72`: The option `--arg-type-hints-in-docstring` is `True` '
-    'but there are no type hints in the docstring arg list ',
     'DOC103: Function `func72`: Docstring arguments are different from function '
     'arguments. (Or could be other formatting issues: '
     'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). Arguments in the function signature but not in the '
@@ -193,9 +185,6 @@ def testClassAttributes(
             'correctly document class attributes.)',
             'DOC101: Method `MyClass2.__init__`: Docstring contains fewer arguments than '
             'in function signature. ',
-            'DOC109: Method `MyClass2.__init__`: The option '
-            '`--arg-type-hints-in-docstring` is `True` but there are no type hints in the '
-            'docstring arg list ',
             'DOC103: Method `MyClass2.__init__`: Docstring arguments are different from '
             'function arguments. (Or could be other formatting issues: '
             'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
@@ -241,9 +230,6 @@ def testClassAttributes(
             'in these args do not match: arg2',
             'DOC101: Method `MyClass2.__init__`: Docstring contains fewer arguments than '
             'in function signature. ',
-            'DOC109: Method `MyClass2.__init__`: The option '
-            '`--arg-type-hints-in-docstring` is `True` but there are no type hints in the '
-            'docstring arg list ',
             'DOC103: Method `MyClass2.__init__`: Docstring arguments are different from '
             'function arguments. (Or could be other formatting issues: '
             'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
@@ -501,8 +487,6 @@ expected_skipCheckingShortDocstrings_False = [
     'but there are no argument type hints in the signature ',
     'DOC107: Function `func1`: The option `--arg-type-hints-in-signature` is `True` '
     'but not all args in the signature have type hints ',
-    'DOC109: Function `func1`: The option `--arg-type-hints-in-docstring` is `True` '
-    'but there are no type hints in the docstring arg list ',
     'DOC103: Function `func1`: Docstring arguments are different from function '
     'arguments. (Or could be other formatting issues: '
     'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). Arguments in the '
@@ -517,8 +501,6 @@ expected_skipCheckingShortDocstrings_False = [
     'but there are no argument type hints in the signature ',
     'DOC107: Function `func2`: The option `--arg-type-hints-in-signature` is `True` '
     'but not all args in the signature have type hints ',
-    'DOC109: Function `func2`: The option `--arg-type-hints-in-docstring` is `True` '
-    'but there are no type hints in the docstring arg list ',
     'DOC103: Function `func2`: Docstring arguments are different from function '
     'arguments. (Or could be other formatting issues: '
     'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). Arguments in the '