+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 # Change Log
 
-## [unpublished]
+## [0.5.2] - 2024-06-26
 
 - Changed
+
   - Pinned to a higher version (0.0.9) of docstring_parser_fork
   - Relaxed class attribute checking logic
     - When a class has no docstring, no DOC6xx violations will be reported
     - When a class has a short docstring (and
       `--skip-checking-short-docstrings`) is set to `True`, no DOC6xx
       violations will be reported
+
+- Full diff
+  - https://github.com/jsh9/pydoclint/compare/0.5.1...0.5.2
 
 ## [0.5.1] - 2024-06-24